requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
numpy>=1.19.5
pandas>=1.3.0
tqdm>=4.62.0
faiss-cpu>=1.7.3
python-multipart>=0.0.5
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
# Using existing torch 2.7.0
transformers>=4.30.0
Pillow>=9.0.0
requests>=2.25.0
open-clip-torch>=2.20.0
# Additional dependencies for visual analysis
torch>=2.0.0
torchvision>=0.15.0
open_clip>=2.20.0
# For image processing
matplotlib>=3.5.0
scikit-image>=0.19.0
# For CLIP-based analysis
clip>=1.0
